requirements.txt
asgiref==3.8.1
click==8.1.8;
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
psycopg2-binary==2.9.10
PyPDF2==3.0.1
sqlparse==0.5.3
typing_extensions==4.13.1
uvicorn==0.34.0
whitenoise==6.9.0
